+++ NEWS.md
@@ -1,3 +1,11 @@
-# smallfactor 0.1.0
+# smallfactor 0.1.1 2021-09-02
+
+* Rename `smallfactor()` to `bytefactor()`
+* Add `bitfactor()`
+* Add tests
+
+* Initial release
+
+# smallfactor 0.1.0  2021-08-27
 
 * Initial release